+++ unicorn_binance_websocket_api/sockets.py
@@ -58,6 +58,7 @@ class BinanceWebSocketApiSocket(object):
         self.output = self.manager.stream_list[self.stream_id]['output']
         self.unicorn_fy = UnicornFy()
         self.exchange = manager.get_exchange()
+        self.loop = None
 
     async def start_socket(self):
         logger.info(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, {str(self.channels)}, "
@@ -70,6 +71,16 @@ class BinanceWebSocketApiSocket(object):
                                                      self.markets,
                                                      symbols=self.symbols) as websocket:
                 self.manager.socket_is_ready[self.stream_id] = True
+                try:
+                    if self.manager.stream_list[self.stream_id]['recent_socket_id'] != self.socket_id:
+                        logger.error(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, "
+                                     f"{str(self.channels)}, {str(self.markets)} socket_id={str(self.socket_id)} "
+                                     f"recent_socket_id={str(self.socket_id)} - Sending payload - exit because its "
+                                     f"not the recent socket id! stream_id={str(self.stream_id)}, recent_socket_id="
+                                     f"{str(self.manager.stream_list[self.stream_id]['recent_socket_id'])}")
+                        sys.exit(0)
+                except KeyError:
+                    sys.exit(1)
                 while self.manager.is_stop_request(self.stream_id) is False and \
                         self.manager.is_stop_as_crash_request(self.stream_id) is False:
                     if self.manager.is_stop_request(self.stream_id):
@@ -78,16 +89,6 @@ class BinanceWebSocketApiSocket(object):
                         sys.exit(0)
                     elif self.manager.is_stop_as_crash_request(self.stream_id):
                         await websocket.close()
-                        sys.exit(1)
-                    try:
-                        if self.manager.stream_list[self.stream_id]['recent_socket_id'] != self.socket_id:
-                            logger.error(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, "
-                                         f"{str(self.channels)}, {str(self.markets)} socket_id={str(self.socket_id)} "
-                                         f"recent_socket_id={str(self.socket_id)} - Sending payload - exit because its "
-                                         f"not the recent socket id! stream_id={str(self.stream_id)}, recent_socket_id="
-                                         f"{str(self.manager.stream_list[self.stream_id]['recent_socket_id'])}")
-                            sys.exit(0)
-                    except KeyError:
                         sys.exit(1)
                     while self.manager.stream_list[self.stream_id]['payload']:
                         logger.info(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, "
@@ -107,44 +108,57 @@ class BinanceWebSocketApiSocket(object):
                         # To avoid a ban we respect the limits of binance:
                         # https://github.com/binance-exchange/binance-official-api-docs/blob/5fccfd572db2f530e25e302c02be5dec12759cf9/CHANGELOG.md#2020-04-23
                         # Limit: max 5 messages per second inclusive pings/pong
-                        max_subscriptions_per_second = self.manager.max_send_messages_per_second - \
-                                                       self.manager.max_send_messages_per_second_reserve
-                        idle_time = 1/max_subscriptions_per_second
-                        time.sleep(idle_time)
+                        max_subscriptions_per_second = self.manager.max_send_messages_per_second - self.manager.max_send_messages_per_second_reserve
+                        idle_time = 1 / max_subscriptions_per_second
+                        await asyncio.sleep(idle_time)
                     try:
-                        received_stream_data_json = await websocket.receive()
+                        received_stream_data_json = await self.loop.create_task(websocket.receive())
                         if received_stream_data_json is not None:
                             if self.output == "UnicornFy":
                                 if self.exchange == "binance.com":
-                                    received_stream_data = self.unicorn_fy.binance_com_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-testnet":
-                                    received_stream_data = self.unicorn_fy.binance_com_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-margin":
-                                    received_stream_data = self.unicorn_fy.binance_com_margin_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_margin_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-margin-testnet":
-                                    received_stream_data = self.unicorn_fy.binance_com_margin_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_margin_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-isolated_margin":
-                                    received_stream_data = self.unicorn_fy.binance_com_isolated_margin_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_isolated_margin_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-isolated_margin-testnet":
-                                    received_stream_data = self.unicorn_fy.binance_com_isolated_margin_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_isolated_margin_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-futures":
-                                    received_stream_data = self.unicorn_fy.binance_com_futures_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_futures_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-futures-testnet":
-                                    received_stream_data = self.unicorn_fy.binance_com_futures_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_futures_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.com-coin-futures" or self.exchange == "binance.com-coin_futures":
-                                    received_stream_data = self.unicorn_fy.binance_com_coin_futures_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_com_coin_futures_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.je":
-                                    received_stream_data = self.unicorn_fy.binance_je_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_je_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.us":
-                                    received_stream_data = self.unicorn_fy.binance_us_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_us_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "trbinance.com":
-                                    received_stream_data = self.unicorn_fy.binance_tr_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_tr_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "jex.com":
                                     received_stream_data = self.unicorn_fy.jex_com_websocket(received_stream_data_json)
                                 elif self.exchange == "binance.org":
-                                    received_stream_data = self.unicorn_fy.binance_org_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_org_websocket(
+                                        received_stream_data_json)
                                 elif self.exchange == "binance.org-testnet":
-                                    received_stream_data = self.unicorn_fy.binance_org_websocket(received_stream_data_json)
+                                    received_stream_data = self.unicorn_fy.binance_org_websocket(
+                                        received_stream_data_json)
                                 else:
                                     received_stream_data = received_stream_data_json
                             elif self.output == "dict":
@@ -170,25 +184,28 @@ class BinanceWebSocketApiSocket(object):
                                     '"code":' in str(received_stream_data_json):
                                 logger.error("BinanceWebSocketApiSocket.start_socket(" +
                                              str(self.stream_id) + ") "
-                                             "- Received error message: " + str(received_stream_data_json))
+                                                                   "- Received error message: " + str(
+                                    received_stream_data_json))
                                 self.manager.add_to_ringbuffer_error(received_stream_data_json)
                             elif '"result":' in str(received_stream_data_json):
                                 logger.info("BinanceWebSocketApiSocket.start_socket(" +
                                             str(self.stream_id) + ") "
-                                            "- Received result message: " + str(received_stream_data_json))
+                                                                  "- Received result message: " + str(
+                                    received_stream_data_json))
                                 self.manager.add_to_ringbuffer_result(received_stream_data_json)
                             else:
                                 if self.manager.stream_list[self.stream_id]['last_received_data_record'] is None:
                                     self.manager.process_stream_signals("FIRST_RECEIVED_DATA",
                                                                         self.stream_id,
                                                                         received_stream_data)
-                                    self.manager.stream_list[self.stream_id]['last_stream_signal'] = "FIRST_RECEIVED_DATA"
-                                self.manager.stream_list[self.stream_id]['last_received_data_record'] = received_stream_data
-
+                                    self.manager.stream_list[self.stream_id][
+                                        'last_stream_signal'] = "FIRST_RECEIVED_DATA"
+                                self.manager.stream_list[self.stream_id][
+                                    'last_received_data_record'] = received_stream_data
                     except websockets.exceptions.ConnectionClosed as error_msg:
                         logger.debug("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
                                      str(self.channels) + ", " + str(self.markets) + ") - Exception ConnectionClosed"
-                                     " - error_msg: " + str(error_msg))
+                                                                                     " - error_msg: " + str(error_msg))
                         if "WebSocket connection is closed: code = 1008" in str(error_msg):
                             websocket.close()
                             self.manager.stream_is_crashing(self.stream_id, error_msg)
@@ -212,14 +229,170 @@ class BinanceWebSocketApiSocket(object):
                     except AttributeError as error_msg:
                         logger.error("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
                                      str(self.channels) + ", " + str(self.markets) + ") - Exception AttributeError - "
-                                     "error_msg: " + str(error_msg))
+                                                                                     "error_msg: " + str(error_msg))
                         self.manager.stream_is_crashing(self.stream_id, str(error_msg))
                         self.manager.set_restart_request(self.stream_id)
                         sys.exit(1)
-                    except SystemExit as error_code:
-                        logger.debug(f"BinanceWebSocketApiSocket.start_socket() (inner) stream_id={self.stream_id} "
-                                     f"- SystemExit({str(error_code)}) - Going to close thread and loop!")
-                        sys.exit(1)
+        except asyncio.TimeoutError as error_msg:
+            self.manager.stream_is_crashing(self.stream_id, error_msg)
+            self.manager.set_restart_request(self.stream_id)
+            sys.exit(1)
+
+    async def send_payload(self, websocket):
+        while self.manager.is_stop_request(self.stream_id) is False and \
+                self.manager.is_stop_as_crash_request(self.stream_id) is False:
+            if self.manager.is_stop_request(self.stream_id):
+                self.manager.stream_is_stopping(self.stream_id)
+                await websocket.close()
+                sys.exit(0)
+            elif self.manager.is_stop_as_crash_request(self.stream_id):
+                await websocket.close()
+                sys.exit(1)
+            while self.manager.stream_list[self.stream_id]['payload']:
+                logger.info(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, "
+                            f"{str(self.channels)}, {str(self.markets)} socket_id={str(self.socket_id)} "
+                            f"recent_socket_id={str(self.socket_id)} - Sending payload started ...")
+                if self.manager.stream_list[self.stream_id]['recent_socket_id'] != self.socket_id:
+                    logger.error(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, "
+                                 f"{str(self.channels)}, {str(self.markets)} socket_id={str(self.socket_id)} "
+                                 f"recent_socket_id={str(self.socket_id)} - Sending payload - exit because its "
+                                 f"not the recent socket id! stream_id={str(self.stream_id)}, recent_socket_id="
+                                 f"{str(self.manager.stream_list[self.stream_id]['recent_socket_id'])}")
+                    sys.exit(0)
+                payload = self.manager.stream_list[self.stream_id]['payload'].pop(0)
+                logger.info(f"BinanceWebSocketApiSocket.start_socket({str(self.stream_id)}, "
+                            f"{str(self.channels)}, {str(self.markets)} - Sending payload: {str(payload)}")
+                await websocket.send(json.dumps(payload, ensure_ascii=False))
+                # To avoid a ban we respect the limits of binance:
+                # https://github.com/binance-exchange/binance-official-api-docs/blob/5fccfd572db2f530e25e302c02be5dec12759cf9/CHANGELOG.md#2020-04-23
+                # Limit: max 5 messages per second inclusive pings/pong
+                max_subscriptions_per_second = self.manager.max_send_messages_per_second - self.manager.max_send_messages_per_second_reserve
+                idle_time = 1 / max_subscriptions_per_second
+                await asyncio.sleep(idle_time)
+
+    async def receive(self, websocket):
+        while self.manager.is_stop_request(self.stream_id) is False and \
+                self.manager.is_stop_as_crash_request(self.stream_id) is False:
+            if self.manager.is_stop_request(self.stream_id):
+                self.manager.stream_is_stopping(self.stream_id)
+                await websocket.close()
+                sys.exit(0)
+            elif self.manager.is_stop_as_crash_request(self.stream_id):
+                await websocket.close()
+                sys.exit(1)
+            try:
+                received_stream_data_json = await websocket.receive()
+                if received_stream_data_json is not None:
+                    if self.output == "UnicornFy":
+                        if self.exchange == "binance.com":
+                            received_stream_data = self.unicorn_fy.binance_com_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.com-testnet":
+                            received_stream_data = self.unicorn_fy.binance_com_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.com-margin":
+                            received_stream_data = self.unicorn_fy.binance_com_margin_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.com-margin-testnet":
+                            received_stream_data = self.unicorn_fy.binance_com_margin_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.com-isolated_margin":
+                            received_stream_data = self.unicorn_fy.binance_com_isolated_margin_websocket(
+                                received_stream_data_json)
+                        elif self.exchange == "binance.com-isolated_margin-testnet":
+                            received_stream_data = self.unicorn_fy.binance_com_isolated_margin_websocket(
+                                received_stream_data_json)
+                        elif self.exchange == "binance.com-futures":
+                            received_stream_data = self.unicorn_fy.binance_com_futures_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.com-futures-testnet":
+                            received_stream_data = self.unicorn_fy.binance_com_futures_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.com-coin-futures" or self.exchange == "binance.com-coin_futures":
+                            received_stream_data = self.unicorn_fy.binance_com_coin_futures_websocket(
+                                received_stream_data_json)
+                        elif self.exchange == "binance.je":
+                            received_stream_data = self.unicorn_fy.binance_je_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.us":
+                            received_stream_data = self.unicorn_fy.binance_us_websocket(received_stream_data_json)
+                        elif self.exchange == "trbinance.com":
+                            received_stream_data = self.unicorn_fy.binance_tr_websocket(received_stream_data_json)
+                        elif self.exchange == "jex.com":
+                            received_stream_data = self.unicorn_fy.jex_com_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.org":
+                            received_stream_data = self.unicorn_fy.binance_org_websocket(received_stream_data_json)
+                        elif self.exchange == "binance.org-testnet":
+                            received_stream_data = self.unicorn_fy.binance_org_websocket(received_stream_data_json)
+                        else:
+                            received_stream_data = received_stream_data_json
+                    elif self.output == "dict":
+                        received_stream_data = json.loads(received_stream_data_json)
+                    else:
+                        received_stream_data = received_stream_data_json
+                    try:
+                        stream_buffer_name = self.manager.stream_list[self.stream_id]['stream_buffer_name']
+                    except KeyError:
+                        stream_buffer_name = False
+                    if stream_buffer_name:
+                        # if create_stream() got a stram_buffer_name -> use it
+                        self.manager.add_to_stream_buffer(received_stream_data,
+                                                          stream_buffer_name=stream_buffer_name)
+                    elif self.manager.specific_process_stream_data[self.stream_id] is not None:
+                        # if create_stream() got a callback function -> use it
+                        self.manager.specific_process_stream_data[self.stream_id](received_stream_data)
+                    else:
+                        # Use the default process_stream_data function provided to/by the manager class
+                        self.manager.process_stream_data(received_stream_data,
+                                                         stream_buffer_name=stream_buffer_name)
+                    if '"error":' in str(received_stream_data_json) or \
+                            '"code":' in str(received_stream_data_json):
+                        logger.error("BinanceWebSocketApiSocket.start_socket(" +
+                                     str(self.stream_id) + ") "
+                                                           "- Received error message: " + str(received_stream_data_json))
+                        self.manager.add_to_ringbuffer_error(received_stream_data_json)
+                    elif '"result":' in str(received_stream_data_json):
+                        logger.info("BinanceWebSocketApiSocket.start_socket(" +
+                                    str(self.stream_id) + ") "
+                                                          "- Received result message: " + str(received_stream_data_json))
+                        self.manager.add_to_ringbuffer_result(received_stream_data_json)
+                    else:
+                        if self.manager.stream_list[self.stream_id]['last_received_data_record'] is None:
+                            self.manager.process_stream_signals("FIRST_RECEIVED_DATA",
+                                                                self.stream_id,
+                                                                received_stream_data)
+                            self.manager.stream_list[self.stream_id]['last_stream_signal'] = "FIRST_RECEIVED_DATA"
+                        self.manager.stream_list[self.stream_id]['last_received_data_record'] = received_stream_data
+            except websockets.exceptions.ConnectionClosed as error_msg:
+                logger.debug("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
+                             str(self.channels) + ", " + str(self.markets) + ") - Exception ConnectionClosed"
+                                                                             " - error_msg: " + str(error_msg))
+                if "WebSocket connection is closed: code = 1008" in str(error_msg):
+                    websocket.close()
+                    self.manager.stream_is_crashing(self.stream_id, error_msg)
+                    self.manager.set_restart_request(self.stream_id)
+                    sys.exit(1)
+                elif "WebSocket connection is closed: code = 1006" in str(error_msg):
+                    self.manager.stream_is_crashing(self.stream_id, error_msg)
+                    self.manager.set_restart_request(self.stream_id)
+                    sys.exit(1)
+                elif "sent 1011 (unexpected error) keepalive ping timeout" in str(error_msg):
+                    self.manager.stream_is_crashing(self.stream_id, error_msg)
+                    self.manager.set_restart_request(self.stream_id)
+                    sys.exit(1)
+                else:
+                    logger.error(f"BinanceWebSocketApiSocket.start_socket({self.stream_id}, {self.channels}, "
+                                 f"{self.markets}) - Unkown exception in ConnectionClosed - error_msg: "
+                                 f"{error_msg}")
+                    self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+                    self.manager.set_restart_request(self.stream_id)
+                    sys.exit(1)
+            except AttributeError as error_msg:
+                logger.error("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
+                             str(self.channels) + ", " + str(self.markets) + ") - Exception AttributeError - "
+                                                                             "error_msg: " + str(error_msg))
+                self.manager.stream_is_crashing(self.stream_id, str(error_msg))
+                self.manager.set_restart_request(self.stream_id)
+                sys.exit(1)
+
+# Todo: Obsolete?
+#                    except SystemExit as error_code:
+#                        logger.debug(f"BinanceWebSocketApiSocket.start_socket() (inner) stream_id={self.stream_id} "
+#                                     f"- SystemExit({str(error_code)}) - Going to close thread and loop!")
+#                        sys.exit(1)
 # Todo: Testing without:
 #                    except KeyError as error_msg:
 #                        logger.error("BinanceWebSocketApiSocket.start_socket(" + str(self.stream_id) + ", " +
@@ -232,12 +405,3 @@ class BinanceWebSocketApiSocket(object):
 #                        self.manager.stream_is_crashing(self.stream_id, str(error_msg))
 #                        self.manager.set_restart_request(self.stream_id)
 #                        sys.exit(1)
-        except asyncio.TimeoutError as error_msg:
-            # Catching https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/issues/221
-            self.manager.stream_is_crashing(self.stream_id, error_msg)
-            self.manager.set_restart_request(self.stream_id)
-            sys.exit(1)
-        except SystemExit as error_code:
-            logger.debug(f"BinanceWebSocketApiSocket.start_socket() (outer) stream_id={self.stream_id} "
-                         f"- SystemExit({str(error_code)}) - Going to close thread and loop!")
-            sys.exit(1)