+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Discussions about unicorn-binance-websocket-api releases!](https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/discussions/categories/releases)
 
 ## 1.41.0.dev (development stage/unreleased/unstable)
+### Added
+- `debug` (bool) parameter to `BinanceWebSocketApiManager()`
+- `manager.get_debug_log()`
+- `manager.set_socket_is_ready()`
+- `manager.set_socket_is_not_ready()`
 
 ## 1.41.0
 ### Added